mingw: work around long path limitations of chdir()
By default, the length of the current directory is limited to 260
characters (or actually a little bit less, for technical reasons).

This limitation is lifted as of Windows 10 1607, but only when
HKLM\SYSTEM\CurrentControlSet\Control\FileSystem LongPathsEnabled is set
to 1 and only when there is a manifest in the .exe file that says that
the application is "longPathAware".

This does not help on earlier Windows versions, or when the user is not
at liberty to change registry settings.

There is another trick up our sleeves, though: short paths. While they
are no longer enabled by default on most drives, they are still enabled
on the C: drive. So at least on the C: drive, we can convert the path
into a short path and still cd to it. Yay.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>