Do use Unix line endings when writing `.config`
We actually use the contents of `.config` in the test suite: if an
old-style test contains a `FEATURE:` line, we specifically check whether
the corresponding feature(s) have been "not set", and we do that by
grepping for an exact line in `.config`.

When using MSYS2's `grep`, the `$` does not actually match DOS line
endings: `\r` is considered a valid non-line-ending character.

So let's make sure to ignore the implicit convention of `fopen()` on
Windows to write DOS line endings, and force them to Unix ones, via the
`b` flag.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>