Fix `xmalloc_readlink()` again
In e86a3ddd8 (win32: make readlink(2) implementation unconditional,
2021-02-12), we removed the special casing of `errno == ENOSYS` when
trying to follow symlinks. However, that handling really was necessary:
when we followed a symlink, and found a non-symlink, and then called
`readlink()` with that non-symlink, we got `errno == ENOSYS` on Windows
(translated from `ERROR_NOT_A_REPARSE_POINT`), and we did want to stop the
loop and return the current path in that case.

To solve that problem, we could revert the part of the patch which
handled `ENOSYS` on Windows. However, let's translate
`ERROR_NOT_A_REPARSE_POINT` to `EINVAL` instead.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>