tests: don't let DEBUG=1 interfere with expected output
A couple of test cases in `tar.tests`, `bc.tests` and `dc.tests` verify
_precise_ output against a known-good version. That version does not
include traced commands, of course.

The symptom looks like this:

	$ SKIP_KNOWN_BUGS=1 DEBUG=1 ../busybox sh ./runtest -v tar
	[...]
	+ eval '\
	exec 2>&1
	uudecode -o input && tar xvf input; echo $?
	ls /tmp/bb_test_evilfile
	ls bb_test_evilfile
	ls symlink/bb_test_evilfile
	'
	+ exec
	+ RETVAL=0
	+ '[' -z 1 ]
	+ set +x
	FAIL: Symlink attack: create symlink and then write through it
	--- expected
	+++ actual
	@@ -1,8 +1,14 @@
	++ uudecode -o input
	++ tar xvf input
	 anything.txt
	 symlink
	 symlink/bb_test_evilfile
	 tar: can't create symlink 'symlink' to '/tmp'
	++ echo 1
	 1
	++ ls /tmp/bb_test_evilfile
	 ls: /tmp/bb_test_evilfile: No such file or directory
	++ ls bb_test_evilfile
	 ls: bb_test_evilfile: No such file or directory
	++ ls symlink/bb_test_evilfile
	 symlink/bb_test_evilfile
	PASS: Symlinks and hardlinks coexist

The only difference between `expect` and `actual` are the traced
commands.

Lets' use `BASH_XTRACEFD` to keep those traced commands out of the
redirected output.

Obviously, this only works if running in a Bash, or when BusyBox is
built with `FEATURE_ASH_BASH_COMPAT`. It won't work otherwise. But
that's still an improvement.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>