process: special-case `cmd //c <...>`
In MSYS2's Bash, this used to be the way to ensure that `/c` is not
converted to `C:\` (in the meantime, `cmd /c <...>` is special-cased
_not_ to do that).

Let's be nice and support that special-case in BusyBox, too. This is
needed to let Git's t3700.56 "can add files via NTFS junctions" pass.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>